requirements.txt
python-dotenv
cassandra-driver